+++ src/cli.ts
@@ -3,7 +3,7 @@ import wtf from "wtfnode";
 
 import chalk from "chalk";
 
-import Args from "./command/arg";
+import {Args, Command} from "./command/arg";
 import { PortsCommand } from "./command/os/ports";
 
 import { ConfigCommand } from "./command/os/config";
@@ -20,8 +20,8 @@ import { LogoutCommand } from "./command/user/logout";
 
 import { OperationInfoCommand } from "./command/operation/info";
 import { OperationListCommand } from "./command/operation/list";
+import {HelpCommand} from "./command/help";
 
-const packageverion = require(`../package.json`).version;
 
 // ========== Global Errors =========
 
@@ -36,7 +36,7 @@ process.on("unhandledRejection", err => {
 });
 
 // ========== Routes =========
-const routes = {
+const routes: Record<string, Command> = {
   signin: {
     help: `Signin to obniz Cloud`,
     async execute(args: any) {
@@ -60,6 +60,7 @@ const routes = {
   "os:config": ConfigCommand,
   "os:config-via-wifi": ConfigViaWifiCommand,
   "os:erase": {
+    help: 'Fully erase a flash on target device.',
     async execute(args: any) {
       const obj = await PreparePort(args);
       obj.stdout = (text: string) => {
@@ -77,41 +78,11 @@ const routes = {
   },
   "operation:list": OperationListCommand,
   "operation:info": OperationInfoCommand,
-  help: async () => {
-    console.log(`
-       _           _               _ _
-  ___ | |__  _ __ (_)____      ___| (_)
- / _ \\| '_ \\| '_ \\| |_  /____ / __| | |
-| (_) | |_) | | | | |/ /_____| (__| | |
- \\___/|_.__/|_| |_|_/___|     \\___|_|_|
-
-
-CLI to interact with obniz device and cloud.
-
-VERSION
-  obniz-cli/${packageverion}
-
-USAGE
-  $ obniz-cli [COMMAND]
-
-COMMANDS
-
-  signin              Signin to obniz cloud.
-  signout             Signout
-
-  user:info           Show current Logged in user
-
-  os:flash-create     Flashing and configure target device and registrate it on your account on obnizCloud.
-  os:flash            Flashing and configure target device.
-  os:config           Configure obnizOS flashed device.
-  os:config-via-wifi  Configure ObnizOS network via Wi-Fi from devices.
-  os:erase            Fully erase a flash on target device.
-  os:list             List of available obnizOS hardwares and versions
-  os:ports            Getting serial ports on your machine.
-
-  operation:list      List of available operations.
-  operation:info      Show operation info.
-  `);
+  help: {
+    help: `Show help`,
+    async execute() {
+      await HelpCommand();
+    }
   }
 };
