+++ src/command/help.ts
@@ -1,0 +1,39 @@
+
+const packageverion = require(`../package.json`).version;
+
+export const HelpCommand = async () => {
+  console.log(`
+       _           _               _ _
+  ___ | |__  _ __ (_)____      ___| (_)
+ / _ \\| '_ \\| '_ \\| |_  /____ / __| | |
+| (_) | |_) | | | | |/ /_____| (__| | |
+ \\___/|_.__/|_| |_|_/___|     \\___|_|_|
+
+
+CLI to interact with obniz device and cloud.
+
+VERSION
+  obniz-cli/${packageverion}
+
+USAGE
+  $ obniz-cli [COMMAND]
+
+COMMANDS
+
+  signin              Signin to obniz cloud.
+  signout             Signout
+
+  user:info           Show current Logged in user
+
+  os:flash-create     Flashing and configure target device and registrate it on your account on obnizCloud.
+  os:flash            Flashing and configure target device.
+  os:config           Configure obnizOS flashed device.
+  os:config-via-wifi  Configure ObnizOS network via Wi-Fi from devices.
+  os:erase            Fully erase a flash on target device.
+  os:list             List of available obnizOS hardwares and versions
+  os:ports            Getting serial ports on your machine.
+
+  operation:list      List of available operations.
+  operation:info      Show operation info.
+  `);
+}
